requirements.txt
moviepy==1.0.3 
numpy==1.23.0
opencv-python==4.9.0.80
soundfile==0.12.1
torch==2.0.1 
torchvision==0.15.2
typing-extensions==4.10.0
PyYAML==6.0.1
openai-whisper==20230314
jiwer==3.0.1
einops==0.7.0 
Cython==3.0.10 
mediapipe==0.10.11
scikit-image==0.21.0
